rename LV_TRACE -> LV_PROFILER
