+++ lv_conf_template.h
@@ -700,11 +700,11 @@
 /*1: Enable system monitor component*/
 #define LV_USE_SYSMON 0
 
-/*1: Enable custom trace system*/
-#define LV_USE_TRACE 0
-#if LV_USE_TRACE
-#define LV_TRACE_INCLUDE <stdint.h>
-#endif
+/*1: Enable the run-time performance profiler*/
+#define LV_USE_PROFILER 0
+#define LV_PROFILER_INCLUDE <stdint.h>
+#define LV_PROFILER_BEGIN
+#define LV_PROFILER_END
 
 /*1: Enable Monkey test*/
 #define LV_USE_MONKEY 0
