+++ src/draw/lv_draw.c
@@ -44,7 +44,9 @@ void lv_draw_init(void)
 
 void lv_draw_wait_for_finish(lv_draw_ctx_t * draw_ctx)
 {
+    LV_PROFILER_BEGIN;
     if(draw_ctx->wait_for_finish) draw_ctx->wait_for_finish(draw_ctx);
+    LV_PROFILER_END;
 }
 
 /**********************
