+++ src/draw/lv_draw_layer.c
@@ -45,6 +45,7 @@ lv_draw_layer_ctx_t * lv_draw_layer_create(lv_draw_ctx_t * draw_ctx, const lv_ar
         LV_LOG_WARN("Couldn't allocate a new layer context");
         return NULL;
     }
+    LV_PROFILER_BEGIN;
 
     lv_memzero(layer_ctx, draw_ctx->layer_instance_size);
 
@@ -55,27 +56,33 @@ lv_draw_layer_ctx_t * lv_draw_layer_create(lv_draw_ctx_t * draw_ctx, const lv_ar
     layer_ctx->area_full = *layer_area;
 
     lv_draw_layer_ctx_t * init_layer_ctx =  draw_ctx->layer_init(draw_ctx, layer_ctx, flags);
+
     if(NULL == init_layer_ctx) {
         lv_free(layer_ctx);
     }
+    LV_PROFILER_END;
     return init_layer_ctx;
 }
 
 void lv_draw_layer_adjust(struct _lv_draw_ctx_t * draw_ctx, struct _lv_draw_layer_ctx_t * layer_ctx,
                           lv_draw_layer_flags_t flags)
 {
+    LV_PROFILER_BEGIN;
     if(draw_ctx->layer_adjust) draw_ctx->layer_adjust(draw_ctx, layer_ctx, flags);
+    LV_PROFILER_END;
 }
 
 void lv_draw_layer_blend(struct _lv_draw_ctx_t * draw_ctx, struct _lv_draw_layer_ctx_t * layer_ctx,
                          lv_draw_img_dsc_t * draw_dsc)
 {
+    LV_PROFILER_BEGIN;
     if(draw_ctx->layer_blend) draw_ctx->layer_blend(draw_ctx, layer_ctx, draw_dsc);
+    LV_PROFILER_END;
 }
 
 void lv_draw_layer_destroy(lv_draw_ctx_t * draw_ctx, lv_draw_layer_ctx_t * layer_ctx)
 {
-
+    LV_PROFILER_BEGIN;
     lv_draw_wait_for_finish(draw_ctx);
     draw_ctx->buf = layer_ctx->original.buf;
     draw_ctx->buf_area = layer_ctx->original.buf_area;
@@ -84,6 +91,7 @@ void lv_draw_layer_destroy(lv_draw_ctx_t * draw_ctx, lv_draw_layer_ctx_t * layer
 
     if(draw_ctx->layer_destroy) draw_ctx->layer_destroy(draw_ctx, layer_ctx);
     lv_free(layer_ctx);
+    LV_PROFILER_END;
 }
 
 /**********************
