+++ src/draw/lv_draw_transform.c
@@ -46,7 +46,9 @@ void lv_draw_transform(lv_draw_ctx_t * draw_ctx, const lv_area_t * dest_area, co
         return;
     }
 
+    LV_PROFILER_BEGIN;
     draw_ctx->draw_transform(draw_ctx, dest_area, src_buf, src_w, src_h, src_stride, draw_dsc, sup, cf, cbuf, abuf);
+    LV_PROFILER_END;
 }
