+++ src/lv_conf_internal.h
@@ -2346,22 +2346,34 @@
     #endif
 #endif
 
-/*1: Enable custom trace system*/
-#ifndef LV_USE_TRACE
-    #ifdef CONFIG_LV_USE_TRACE
-        #define LV_USE_TRACE CONFIG_LV_USE_TRACE
+/*1: Enable the run-time performance profiler*/
+#ifndef LV_USE_PROFILER
+    #ifdef CONFIG_LV_USE_PROFILER
+        #define LV_USE_PROFILER CONFIG_LV_USE_PROFILER
     #else
-        #define LV_USE_TRACE 0
+        #define LV_USE_PROFILER 0
     #endif
 #endif
-#if LV_USE_TRACE
-#ifndef LV_TRACE_INCLUDE
-    #ifdef CONFIG_LV_TRACE_INCLUDE
-        #define LV_TRACE_INCLUDE CONFIG_LV_TRACE_INCLUDE
+#ifndef LV_PROFILER_INCLUDE
+    #ifdef CONFIG_LV_PROFILER_INCLUDE
+        #define LV_PROFILER_INCLUDE CONFIG_LV_PROFILER_INCLUDE
     #else
-        #define LV_TRACE_INCLUDE <stdint.h>
+        #define LV_PROFILER_INCLUDE <stdint.h>
     #endif
 #endif
+#ifndef LV_PROFILER_BEGIN
+    #ifdef CONFIG_LV_PROFILER_BEGIN
+        #define LV_PROFILER_BEGIN CONFIG_LV_PROFILER_BEGIN
+    #else
+        #define LV_PROFILER_BEGIN
+    #endif
+#endif
+#ifndef LV_PROFILER_END
+    #ifdef CONFIG_LV_PROFILER_END
+        #define LV_PROFILER_END CONFIG_LV_PROFILER_END
+    #else
+        #define LV_PROFILER_END
+    #endif
 #endif
 
 /*1: Enable Monkey test*/
